+++ service/updateAvatar.js
